requirements.txt
numpy==1.19.2
matplotlib==3.3.4
scikit-learn==0.24.1
keras==2.4.3
opencv-python==4.5.1.48
tensorflow==2.4.1
tensorflow-gpu==2.4.1
Pillow==8.1.2